# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TTLG Backend",
    "description": "HTTP API ядра системы сопровождения учебного процесса: пользователи, занятия, домашние задания, диалог с ассистентом (LLM). Версия в пути: `/v1`.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Создать пользователя",
        "description": "Регистрирует пользователя (MVP). В будущем потребуется реальная авторизация.",
        "operationId": "create_user_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Получить пользователя по id",
        "description": "Возвращает карточку пользователя или 404.",
        "operationId": "read_user_v1_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}/progress": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Сводка прогресса ученика",
        "description": "Агрегаты по занятиям и домашним заданиям для указанного пользователя.",
        "operationId": "read_user_progress_v1_users__user_id__progress_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressSummaryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressSummaryRead": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id",
            "description": "UUID ученика строкой"
          },
          "lessons_completed": {
            "type": "integer",
            "title": "Lessons Completed",
            "description": "Число завершённых занятий"
          },
          "lessons_total": {
            "type": "integer",
            "title": "Lessons Total",
            "description": "Всего занятий у ученика"
          },
          "assignments_done": {
            "type": "integer",
            "title": "Assignments Done",
            "description": "Число выполненных ДЗ"
          },
          "assignments_total": {
            "type": "integer",
            "title": "Assignments Total",
            "description": "Всего назначенных ДЗ"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "lessons_completed",
          "lessons_total",
          "assignments_done",
          "assignments_total"
        ],
        "title": "ProgressSummaryRead"
      },
      "UserCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Отображаемое имя пользователя"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Роль: ученик или преподаватель"
          },
          "telegram_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Id",
            "description": "Telegram user id; для бота обязателен у ученика"
          }
        },
        "type": "object",
        "required": [
          "name",
          "role"
        ],
        "title": "UserCreate"
      },
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Внутренний идентификатор пользователя"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Имя"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole",
            "description": "Роль"
          },
          "telegram_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Id",
            "description": "Привязка к Telegram, если задана"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Время создания записи (UTC)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "role",
          "created_at"
        ],
        "title": "UserRead"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "student",
          "teacher"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}